reset buffers of LMs and SMs
